@@ -10,30 +10,36 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.4",
         "ewallet/application": "1.0",
-        "monolog/monolog": "1.25.3",
-        "slim/slim": "3.12.3",
-        "twig/twig": "2.12.3",
-        "vlucas/phpdotenv": "3.6.0"
+        "guzzlehttp/psr7": "1.7.0",
+        "http-interop/http-factory-guzzle": "1.0.0",
+        "monolog/monolog": "2.2.0",
+        "slim/slim": "4.7.1",
+        "symfony/dependency-injection": "5.2.1",
+        "teapot/status-code": "1.1.2",
+        "twig/twig": "3.1.1",
+        "vlucas/phpdotenv": "5.2.0"
     },
     "require-dev": {
-        "codeception/codeception": "4.0.1",
-        "codeception/module-webdriver": "1.0.0",
-        "consolidation/robo": "1.4.11",
-        "friendsofphp/php-cs-fixer": "2.16.1",
-        "infection/infection": "0.13.6",
-        "mockery/mockery": "1.3.1",
-        "nelmio/alice": "3.5.8",
-        "phpstan/phpstan": "0.11.19",
-        "phpstan/phpstan-strict-rules": "0.11.1",
-        "phpunit/phpunit": "8.5.1",
-        "roave/security-advisories": "dev-master"
+        "codeception/codeception": "4.1.12",
+        "codeception/module-webdriver": "1.1.4",
+        "friendsofphp/php-cs-fixer": "2.17.2",
+        "infection/infection": "0.20.2",
+        "mockery/mockery": "1.4.2",
+        "nelmio/alice": "3.7.4",
+        "phpspec/phpspec": "^7.0",
+        "phpspec/prophecy-phpunit": "^2.0",
+        "phpstan/phpstan": "0.12.63",
+        "phpstan/phpstan-strict-rules": "0.12.7",
+        "phpunit/phpunit": "9.5.0",
+        "roave/security-advisories": "dev-master",
+        "symfony/finder": "5.2.1"
     },
     "config": {
         "bin-dir": "bin/",
         "platform": {
-            "php": "7.2.19"
+            "php": "7.4.12"
         },
         "sort-packages": true
     },
@@ -42,7 +48,8 @@
             "Application\\": "src/",
             "Ewallet\\": "src/",
             "Adapters": "src/",
-            "UI\\": "src/"
+            "UI\\": "src/",
+            "Framework\\": "src/"
         }
     },
     "autoload-dev": {
