@@ -7,77 +7,57 @@
 
 namespace UI\Slim\Controllers;
 
-use Application\Actions\InputValidator;
+use Application\DomainException;
 use Application\Templating\TemplateEngine;
 use Ewallet\ManageWallet\TransferFunds\TransferFundsAction;
 use Ewallet\ManageWallet\TransferFunds\TransferFundsInput;
-use Ewallet\ManageWallet\TransferFunds\TransferFundsResponder;
-use Ewallet\ManageWallet\TransferFunds\TransferFundsSummary;
 use Ewallet\Memberships\InsufficientFunds;
+use Ewallet\Memberships\InvalidTransfer;
 use Ewallet\Memberships\MemberId;
 use Ewallet\Memberships\MembersWebRepository;
 use Ewallet\Memberships\UnknownMember;
 use Psr\Http\Message\ResponseInterface;
-use Slim\Http\Request;
-use Slim\Http\Response;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+use UI\Slim\ResponseFactory;
 
-class TransferFundsController implements TransferFundsResponder
+final class TransferFundsController implements RequestHandlerInterface
 {
-    /** @var TransferFundsAction */
-    private $action;
+    private TransferFundsAction $action;
 
-    /** @var TemplateEngine */
-    private $template;
+    private TemplateEngine $template;
 
-    /** @var ResponseInterface */
-    private $response;
+    private ResponseFactory $response;
 
-    /** @var MembersWebRepository */
-    private $members;
+    private MembersWebRepository $members;
 
-    public function __construct(TransferFundsAction $action, TemplateEngine $template, MembersWebRepository $members)
-    {
+    public function __construct(
+        TransferFundsAction $action,
+        TemplateEngine $template,
+        MembersWebRepository $members,
+        ResponseFactory $response
+    ) {
         $this->action = $action;
-        $this->action->attach($this);
         $this->template = $template;
         $this->members = $members;
+        $this->response = $response;
     }
 
     /**
      * Perform the transfer
      *
-     * @throws \Ewallet\Memberships\InsufficientFunds If the sender does not
-     * have sufficient funds
-     * @throws \Ewallet\Memberships\InvalidTransfer If the sender tries to
-     * transfer a negative amount
-     * @throws \Ewallet\Memberships\UnknownMember If either the sender or the
-     * recipient are unknown
+     * @throws InsufficientFunds If the sender does not have sufficient funds
+     * @throws InvalidTransfer If the sender tries to transfer a negative amount
+     * @throws UnknownMember If either the sender or the recipient are unknown
      */
-    public function transfer(Request $request): ResponseInterface
+    public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        $this->action->transfer(TransferFundsInput::from((array) $request->getParsedBody()));
+        try {
+            $summary = $this->action->transfer(new TransferFundsInput((array) $request->getParsedBody()));
+        } catch (DomainException $exception) {
+            return $this->respondToDomainException($exception);
+        }
 
-        return $this->response;
-    }
-
-    public function respondToInvalidInput(InputValidator $input): void
-    {
-        $recipients = $this->members->excluding(new MemberId($input->values()['senderId']));
-        $html = $this->template->render('member/transfer-funds.html', [
-            'senderId' => $input->values()['senderId'],
-            'recipientId' => new MemberId($input->values()['recipientId']),
-            'recipients' => $recipients,
-            'amount' => $input->values()['amount'],
-            'errors' => $input->errors(),
-        ]);
-
-        $response = new Response();
-        $response->getBody()->write($html);
-        $this->response = $response;
-    }
-
-    public function respondToTransferCompleted(TransferFundsSummary $summary): void
-    {
         $recipients = $this->members->excluding($summary->senderId());
         $html = $this->template->render('member/transfer-funds.html', [
             'summary' => $summary,
@@ -86,12 +66,10 @@ class TransferFundsController implements TransferFundsResponder
             'recipients' => $recipients,
         ]);
 
-        $response = new Response();
-        $response->getBody()->write($html);
-        $this->response = $response;
+        return $this->response->html($html);
     }
 
-    public function respondToUnknownMember(UnknownMember $exception): void
+    public function respondToDomainException(DomainException $exception): ResponseInterface
     {
         $senderId = new MemberId('ABC');
         $recipients = $this->members->excluding($senderId);
@@ -101,23 +79,6 @@ class TransferFundsController implements TransferFundsResponder
             'errors' => ['senderId' => $exception->getMessage()],
         ]);
 
-        $response = new Response();
-        $response->getBody()->write($html);
-        $this->response = $response;
-    }
-
-    public function respondToInsufficientFunds(InsufficientFunds $exception): void
-    {
-        $senderId = new MemberId('ABC');
-        $recipients = $this->members->excluding($senderId);
-        $html = $this->template->render('member/transfer-funds.html', [
-            'senderId' => $senderId,
-            'recipients' => $recipients,
-            'errors' => ['amount' => $exception->getMessage()],
-        ]);
-
-        $response = new Response();
-        $response->getBody()->write($html);
-        $this->response = $response;
+        return $this->response->html($html);
     }
 }
