@@ -7,15 +7,18 @@
 
 require __DIR__ . '/../vendor/autoload.php';
 
+use Adapters\Symfony\DependencyInjection\ContainerFactory;
 use Dotenv\Dotenv;
-use UI\Slim\Application;
+use Framework\Slim\ApplicationFactory;
 
-if (!isset($_ENV['APP_ENV'])) {
-    // We' re not running the application from the containers
-    $environment = Dotenv::create(__DIR__ . '/../');
-    $environment->load();
-    $environment->required(['APP_ENV', 'DB_URL', 'PDO_DRIVER']);
-}
+(static function () {
+    if (!isset($_ENV['APP_ENV'])) {
+        // We' re not running the application from the containers
+        $environment = Dotenv::createImmutable(__DIR__ . '/../');
+        $environment->load();
+        $environment->required(['APP_ENV', 'DB_URL', 'PDO_DRIVER']);
+    }
 
-$app = new Application(require __DIR__ . '/../config.php');
-$app->run();
+    $app = ApplicationFactory::createFromContainer(ContainerFactory::new());
+    $app->run();
+})();
