@@ -6,12 +6,12 @@
  */
 namespace Page;
 
-class TransferFundsPage
+final class TransferFundsPage
 {
-    public static $inputTransferInformationPage = 'transfer-form';
-    public static $transferCompletedPage = 'transfer-funds';
-    public static $recipients = 'Transfer to';
-    public static $amount = 'Amount';
-    public static $transferButton = 'Transfer';
-    public static $successMessage = '#transfer-message';
+    public static string $inputTransferInformationPage = 'transfer-form';
+    public static string $transferCompletedPage = 'transfer-funds';
+    public static string $recipients = 'Transfer to';
+    public static string $amount = 'Amount';
+    public static string $transferButton = 'Transfer';
+    public static string $successMessage = '#transfer-message';
 }
